analytics/requirements.txt
kafka-python==2.0.2
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
